Применил акцентную типографию
